+++ CHANGELOG.md
@@ -6,6 +6,7 @@ From v3.0.2 to v3.0.3
 
 - Fixed a bug where an attached standardizer would be refit when calling 
 `QSPRModel.predictMols` with `use_applicability_domain=True`.
+- Fixed random seed not set in `FoldsFromDataSplit.iterFolds` for `ClusterSplit`.
 
 ## Changes