+++ CHANGELOG.md
@@ -4,8 +4,8 @@ From v3.0.2 to v3.0.3
 
 ## Fixes
 
-- Fixed a bug where an attached standardizer would be refit when calling 
-`QSPRModel.predictMols` with `use_applicability_domain=True`.
+- Fixed a bug where an attached standardizer would be refit when calling
+  `QSPRModel.predictMols` with `use_applicability_domain=True`.
 - Fixed random seed not set in `FoldsFromDataSplit.iterFolds` for `ClusterSplit`.
 
 ## Changes
@@ -14,7 +14,8 @@ None.
 
 ## New Features
 
-None.
+- Added the `prepMols` method to `DescriptorSet` to allow separated customization of
+  molecule preparation before descriptor calculation.
 
 ## Removed Features