+++ CHANGELOG.md
@@ -9,6 +9,8 @@ From v3.0.2 to v3.0.3
 - Fixed a bug with `use_applicability_domain=True` in `QSPRModel.predictMols` 
   where an error would be raised if there were invalid molecules in the input.
 - Fixed a bug where dataset type was not properly set to numeric in `MlChemADWrapper.contains`
+- Fixed a bug where an attached standardizer would be refit when calling
+  `QSPRModel.predictMols` with `use_applicability_domain=True`.
 - Fixed random seed not set in `FoldsFromDataSplit.iterFolds` for `ClusterSplit`.
 
 ## Changes
@@ -17,7 +19,8 @@ From v3.0.2 to v3.0.3
 
 ## New Features
 
-None.
+- Added the `prepMols` method to `DescriptorSet` to allow separated customization of
+  molecule preparation before descriptor calculation.
 
 ## Removed Features