+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-From v3.0.2 to v3.0.3
+From v3.0.1 to v3.0.2
 
 ## Fixes
 
@@ -9,9 +9,11 @@ From v3.0.2 to v3.0.3
 - Fixed a bug with `use_applicability_domain=True` in `QSPRModel.predictMols` 
   where an error would be raised if there were invalid molecules in the input.
 - Fixed a bug where dataset type was not properly set to numeric in `MlChemADWrapper.contains`
+- Fixed a bug in `QSPRDataset` where property transformations were not applied.
 - Fixed a bug where an attached standardizer would be refit when calling
   `QSPRModel.predictMols` with `use_applicability_domain=True`.
 - Fixed random seed not set in `FoldsFromDataSplit.iterFolds` for `ClusterSplit`.
+- Fixed a bug where class ratios were shuffled in the `RatioDistributionAlgorithm`.
 
 ## Changes
 
@@ -19,6 +21,8 @@ From v3.0.2 to v3.0.3
 
 ## New Features
 
+- The `DataFrameDescriptorSet` class was extended to allow more flexibility when joining
+  custom descriptor sets.
 - Added the `prepMols` method to `DescriptorSet` to allow separated customization of
   molecule preparation before descriptor calculation.